requirements.txt
requests
beautifulsoup4
google-generativeai
python-telegram-bot==13.15
pytz
nest_asyncio